Move back to test_mvcc.rs
