+++ storage/storage.rs
@@ -629,136 +629,3 @@ impl StorageOperation {
         StorageOperation::BYTES
     }
 }
-
-#[cfg(test)]
-pub mod tests {
-    use std::path::Path;
-
-    use bytes::byte_array::ByteArray;
-    use bytes::byte_reference::ByteReference;
-    use durability::wal::WAL;
-    use crate::key_value::{StorageKey, StorageKeyArray, StorageKeyReference};
-    use crate::MVCCStorage;
-    use crate::KeyspaceSet;
-    use test_utils::{create_tmp_dir, init_logging};
-
-    macro_rules! test_keyspace_set {
-        {$($variant:ident => $id:literal : $name: literal),* $(,)?} => {
-            #[derive(Copy, Clone)]
-            enum TestKeyspaceSet { $($variant),* }
-            impl KeyspaceSet for TestKeyspaceSet {
-                fn iter() -> impl Iterator<Item = Self> { [$(Self::$variant),*].into_iter() }
-                fn id(&self) -> u8 {
-                    match *self { $(Self::$variant => $id),* }
-                }
-                fn name(&self) -> &'static str {
-                    match *self { $(Self::$variant => $name),* }
-                }
-            }
-        };
-    }
-
-    test_keyspace_set! {
-        Keyspace => 0: "keyspace",
-    }
-
-    use self::TestKeyspaceSet::Keyspace;
-
-    const KEY_1: [u8; 4] = [0x0, 0x0, 0x0, 0x1];
-    const KEY_2: [u8; 4] = [0x0, 0x0, 0x0, 0x2];
-    const VALUE_0: [u8; 1] = [0x0];
-    const VALUE_1: [u8; 1] = [0x1];
-    const VALUE_2: [u8; 1] = [0x2];
-
-    fn setup_storage(storage_path: &Path) -> MVCCStorage<WAL> {
-        let storage = MVCCStorage::recover::<TestKeyspaceSet>("storage", storage_path).unwrap();
-        storage
-    }
-
-    #[test]
-    fn test_commit_increments_watermark() {
-        init_logging();
-        let storage_path = create_tmp_dir();
-        let storage = setup_storage(&storage_path);
-        let wm_initial = storage.isolation_manager.watermark();
-        let snapshot_0 = storage.open_snapshot_write();
-        let result_0 = snapshot_0.commit().unwrap();
-
-        assert_eq!(wm_initial.next().number().number() + 1, storage.isolation_manager.watermark().number().number());
-    }
-
-    #[test]
-    fn test_reading_snapshots() {
-        init_logging();
-        let storage_path = create_tmp_dir();
-        let storage = setup_storage(&storage_path);
-
-        let key_1: &StorageKey<'_, 48> = &StorageKey::Reference(StorageKeyReference::new(Keyspace, ByteReference::new(&KEY_1)));
-
-        let snapshot_write_0 = storage.open_snapshot_write();
-        snapshot_write_0.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_0)).unwrap();
-        snapshot_write_0.commit().unwrap();
-
-        let watermark_0 = storage.isolation_manager.watermark();
-
-        let snapshot_read_0 = storage.open_snapshot_read();
-        assert_eq!(snapshot_read_0.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_0);
-
-        let snapshot_write_1 = storage.open_snapshot_write();
-        snapshot_write_1.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_1)).unwrap();
-        let snapshot_read_01 = storage.open_snapshot_read();
-        assert_eq!(snapshot_read_0.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_0);
-        assert_eq!(snapshot_read_01.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_0);
-
-        let result_write_1 = snapshot_write_1.commit();
-        assert!(result_write_1.is_ok());
-
-        let snapshot_read_1 = storage.open_snapshot_read();
-        assert_eq!(snapshot_read_0.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_0);
-        assert_eq!(snapshot_read_01.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_0);
-        assert_eq!(snapshot_read_1.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_1);
-        snapshot_read_1.close_resources();
-        snapshot_read_01.close_resources();
-        snapshot_read_0.close_resources();
-
-        // Read from further in the past.
-        let snapshot_read_02 = storage.open_snapshot_read_at(watermark_0);
-        assert_eq!(snapshot_read_02.get::<128>(key_1.as_reference()).unwrap().unwrap().bytes(), VALUE_0);
-        snapshot_read_02.close_resources();
-    }
-
-    #[test]
-    fn test_writing_same_key_conflicts() {
-        // TODO: Why does this exist if we have separate isolation tests?
-        init_logging();
-        let storage_path = create_tmp_dir();
-        let storage = setup_storage(&storage_path);
-
-        let key_1: &StorageKey<'_, 48> = &StorageKey::Reference(StorageKeyReference::new(Keyspace, ByteReference::new(&KEY_1)));
-        let key_1: &StorageKey<'_, 48> = &StorageKey::Reference(StorageKeyReference::new(Keyspace, ByteReference::new(&KEY_2)));
-
-        let snapshot_write_0 = storage.open_snapshot_write();
-        snapshot_write_0.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_0)).unwrap();
-        snapshot_write_0.commit().unwrap();
-
-        let snapshot_write_11 = storage.open_snapshot_write();
-        let snapshot_write_21 = storage.open_snapshot_write();
-        snapshot_write_11.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_1)).unwrap();
-        snapshot_write_21.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_1)), ByteArray::copy(&VALUE_2)).unwrap();
-        let result_write_11 = snapshot_write_11.commit();
-        assert!(result_write_11.is_ok());
-        let result_write_21 = snapshot_write_21.commit();
-        assert!(!result_write_21.is_ok()); // Fails
-
-        let snapshot_write_12 = storage.open_snapshot_write();
-        let snapshot_write_22 = storage.open_snapshot_write();
-        snapshot_write_12.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_2)), ByteArray::copy(&VALUE_1)).unwrap();
-        snapshot_write_22.put_val(StorageKeyArray::new(Keyspace, ByteArray::copy(&KEY_2)), ByteArray::copy(&VALUE_2)).unwrap();
-
-        let result_write_22 = snapshot_write_22.commit();
-        assert!(result_write_22.is_ok());
-        let result_write_12 = snapshot_write_12.commit();
-
-        assert!(!result_write_12.is_ok()); // Fail
-    }
-}
